refactor BaseTest classes
