+++ src/test/java/iteration2/api/BaseAPITest.java
@@ -1,0 +1,6 @@
+package iteration2.api;
+
+import iteration2.BaseTest;
+
+public class BaseAPITest extends BaseTest {
+}
