+++ src/test/java/iteration2/api/TransferMoneyTest.java
@@ -17,7 +17,7 @@ import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-public class TransferMoneyTest extends BaseTest {
+public class TransferMoneyTest extends BaseAPITest {
     private static User firstUser;
     private static User secondUser;
