+++ src/test/java/iteration2/ui/BaseUITest.java
@@ -1,0 +1,6 @@
+package iteration2.ui;
+
+import iteration2.BaseTest;
+
+public class BaseUITest extends BaseTest {
+}
